+++ lib/edgedb/protocol/codecs/config_memory.ex
@@ -36,7 +36,7 @@ defimpl EdgeDB.Protocol.Codec, for: EdgeDB.Protocol.Codecs.ConfigMemory do
   @impl EdgeDB.Protocol.Codec
   def encode(_codec, bytes, _codec_storage)
       when is_integer(bytes) and bytes in -0x8000000000000000..0x7FFFFFFFFFFFFFFF do
-    <<8::uint32, bytes::int64>>
+    <<8::uint32(), bytes::int64()>>
   end
 
   @impl EdgeDB.Protocol.Codec
@@ -47,7 +47,7 @@ defimpl EdgeDB.Protocol.Codec, for: EdgeDB.Protocol.Codecs.ConfigMemory do
   end
 
   @impl EdgeDB.Protocol.Codec
-  def decode(_codec, <<8::uint32, bytes::int64>>, _codec_storage) do
+  def decode(_codec, <<8::uint32(), bytes::int64()>>, _codec_storage) do
     %EdgeDB.ConfigMemory{bytes: bytes}
   end
 end