+++ lib/edgedb/protocol/codecs/tuple.ex
@@ -45,18 +45,22 @@ defimpl EdgeDB.Protocol.Codec, for: EdgeDB.Protocol.Codecs.Tuple do
       |> Enum.zip(codecs)
       |> Enum.map(fn
         {nil, _codec} ->
-          <<0::int32, -1::int32>>
+          <<0::int32(), -1::int32()>>
 
         {value, codec} ->
-          [<<0::int32>> | Codec.encode(codec, value, codec_storage)]
+          [<<0::int32()>> | Codec.encode(codec, value, codec_storage)]
       end)
 
-    data = [<<length(values)::int32>> | values]
-    [<<IO.iodata_length(data)::uint32>> | data]
+    data = [<<length(values)::int32()>> | values]
+    [<<IO.iodata_length(data)::uint32()>> | data]
   end
 
   @impl Codec
-  def decode(%{codecs: codecs}, <<_length::uint32, nelems::int32, data::binary>>, codec_storage) do
+  def decode(
+        %{codecs: codecs},
+        <<_length::uint32(), nelems::int32(), data::binary>>,
+        codec_storage
+      ) do
     codecs = Enum.map(codecs, &CodecStorage.get(codec_storage, &1))
     elements = decode_element_list(data, codecs, codec_storage, nelems, [])
     List.to_tuple(elements)
@@ -67,7 +71,7 @@ defimpl EdgeDB.Protocol.Codec, for: EdgeDB.Protocol.Codecs.Tuple do
   end
 
   defp decode_element_list(
-         <<_reserved::int32, -1::int32, rest::binary>>,
+         <<_reserved::int32(), -1::int32(), rest::binary>>,
          [_codec | codecs],
          codec_storage,
          count,
@@ -77,13 +81,13 @@ defimpl EdgeDB.Protocol.Codec, for: EdgeDB.Protocol.Codecs.Tuple do
   end
 
   defp decode_element_list(
-         <<_reserved::int32, length::int32, data::binary(length), rest::binary>>,
+         <<_reserved::int32(), length::int32(), data::binary(length), rest::binary>>,
          [codec | codecs],
          codec_storage,
          count,
          acc
        ) do
-    element = Codec.decode(codec, <<length::int32, data::binary>>, codec_storage)
+    element = Codec.decode(codec, <<length::int32(), data::binary>>, codec_storage)
     decode_element_list(rest, codecs, codec_storage, count - 1, [element | acc])
   end
 end