+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - support for `cal::date_duration` EdgeDB type via `EdgeDB.DateDuration` structure.
 
+### Changed
+
+- implementation of `Enumerable` protocol for `EdgeDB.Set`.
+
 ## [0.5.1] - 2022-08-25
 
 [Compare with 0.5.0](https://github.com/nsidnev/edgedb-elixir/compare/v0.5.0...v0.5.1)